Homework_templated_separate
