+++ FunctionTemplates/FillUniqueRand.h
@@ -1,0 +1,13 @@
+#pragma once
+#include"constans.h"
+#include"stdafx.h"
+void FillRand(int arr[], const int n);
+void FillRand(double arr[], const int n);
+void FillRand(int arr[ROWS][COLS], const int ROWS, const int COLS);
+void FillRand(double arr[ROWS][COLS], const int ROWS, const int COLS);
+void FillRand(char arr[ROWS][COLS], const int ROWS, const int COLS);
+
+void UniqueRand(int arr[], const int n);
+void UniqueRand(int arr[ROWS][COLS], const int ROWS, const int COLS);
+void UniqueRand(double arr[ROWS][COLS], const int ROWS, const int COLS);
+void UniqueRand(char arr[ROWS][COLS], const int ROWS, const int COLS);
