+++ FunctionTemplates/Print.cpp
@@ -1,0 +1,21 @@
+#include"Print.h"
+template<typename T>void Print(const T arr[], const int n)
+{
+	//Вывод массива на экран:
+	for (int i = 0; i < n; i++)
+	{
+		cout << arr[i] << "\t";
+	}
+	cout << endl;
+}
+template<typename T>void Print(const T arr[ROWS][COLS], const int ROWS, const int COLS)
+{
+	for (int i = 0; i < ROWS; i++)
+	{
+		for (int j = 0; j < COLS; j++)
+		{
+			cout << arr[i][j] << "\t";
+		}
+		cout << endl;
+	}
+}
