+++ FunctionTemplates/Sort.h
@@ -1,0 +1,5 @@
+#pragma once
+#include"constans.h"
+#include"stdafx.h"
+template<typename T> void Sort(T arr[], const int n);
+template<typename T> void Sort(T arr[ROWS][COLS], const int ROWS, const int COLS);
