+++ FunctionTemplates/Sum.cpp
@@ -1,0 +1,23 @@
+#include"Sum.h"
+template<typename T>T Sum(const T arr[], const int n)
+{
+	//Вычисление суммы элементов массива:
+	T sum = 0;
+	for (int i = 0; i < n; i++)
+	{
+		sum += arr[i];
+	}
+	return sum;
+}
+template<typename T> T Sum(const T arr[ROWS][COLS], const int ROWS, const int COLS)
+{
+	T sum = 0;
+	for (int i = 0; i < ROWS; i++)
+	{
+		for (int j = 0; j < COLS; j++)
+		{
+			sum += arr[i][j];
+		}
+	}
+	return sum;
+}
