+++ FunctionTemplates/constans.h
@@ -1,0 +1,3 @@
+#pragma once
+const int ROWS = 3;
+const int COLS = 4;
