+++ FunctionTemplates/main.cpp
@@ -1,50 +1,16 @@
-﻿#include<iostream>
-using namespace std;
-
-const int ROWS = 3;
-const int COLS = 4;
-
-void FillRand(int arr[], const int n);
-void FillRand(double arr[], const int n);
-void FillRand(int arr[ROWS][COLS], const int ROWS, const int COLS);
-void FillRand(double arr[ROWS][COLS], const int ROWS, const int COLS);
-void FillRand(char arr[ROWS][COLS], const int ROWS, const int COLS);
-
-template<typename T>void Print(const T arr[], const int n);
-template<typename T>void Print(const T arr[ROWS][COLS], const int ROWS, const int COLS);
-
-template<typename T>T Sum(const T arr[], const int n);
-template<typename T> T Sum(const T arr[ROWS][COLS], const int ROWS, const int COLS);
-
-template<typename T>double Avg(const T arr[], const int n);
-template<typename T> double Avg(const T arr[ROWS][COLS], const int ROWS, const int COLS);
-
-template<typename T>T minValueIn(const T arr[], const int n);
-template<typename T> T minValueIn(const T arr[ROWS][COLS], const int ROWS, const int COLS);
-
-template<typename T>T maxValueIn(const T arr[], const int n);
-template<typename T> T maxValueIn(const T arr[ROWS][COLS], const int ROWS, const int COLS);
-
-template<typename T> void Sort(T arr[], const int n);
-template<typename T> void Sort(T arr[ROWS][COLS], const int ROWS, const int COLS);
-
-template<typename T>void shiftLeft(T arr[], const int n, int number_of_shifts);
-template<typename T> void shiftLeft(T arr[ROWS][COLS], const int ROWS, const int COLS, int number_of_shifts);
-
-template<typename T>void shiftRight(T arr[], const int n, int number_of_shifts);
-template<typename T> void shiftRight(T arr[ROWS][COLS], const int ROWS, const int COLS, int number_of_shifts);
-
-void UniqueRand(int arr[], const int n);
-void UniqueRand(int arr[ROWS][COLS], const int ROWS, const int COLS);
-void UniqueRand(double arr[ROWS][COLS], const int ROWS, const int COLS);
-void UniqueRand(char arr[ROWS][COLS], const int ROWS, const int COLS);
-
-template<typename T> void Search(T arr[], const int n);
-template<typename T> void Search(T arr[ROWS][COLS], const int ROWS, const int COLS);
+﻿#include"stdafx.h"
+#include"constans.h"
+#include"FillUniqueRand.h"
+#include"Print.cpp"
+#include"Sum.cpp"
+#include"Avg.cpp"
+#include"minmaxValueIn.cpp"
+#include"Sort.cpp"
+#include"shiftLeftRight.cpp"
+#include"Search.cpp"
 void main()
 {
 	setlocale(LC_ALL, "");
-
 	const int n = 10;
 	int arr[n] = { 0, 1, 2, 3, 4, 5, 6, 7, 8, 9 };
 	FillRand(arr, n);
@@ -126,386 +92,4 @@ void main()
 	Search(i_arr_2, ROWS, COLS); cout << endl;
 	Search(d_arr_2, ROWS, COLS); cout << endl;
 	Search(c_arr_2, ROWS, COLS); cout << endl;
-}
-void FillRand(int arr[], const int n)
-{
-	//rand();	//возвращает псевдослучайное число в диапазоне от 0 до 32 767 (MAX_RAND)
-	for (int i = 0; i < n; i++)
-	{
-		arr[i] = rand() % 100;
-	}
-}
-void FillRand(double arr[], const int n)
-{
-	//rand();	//возвращает псевдослучайное число в диапазоне от 0 до 32 767 (MAX_RAND)
-	for (int i = 0; i < n; i++)
-	{
-		arr[i] = rand() % 10000;
-		arr[i] /= 100;
-	}
-}
-void FillRand(int arr[ROWS][COLS], const int ROWS, const int COLS)
-{
-	for (int i = 0; i < ROWS; i++)
-	{
-		for (int j = 0; j < COLS; j++)
-		{
-			arr[i][j] = rand() % 15;
-		}
-	}
-}
-void FillRand(double arr[ROWS][COLS], const int ROWS, const int COLS)
-{
-	for (int i = 0; i < ROWS; i++)
-	{
-		for (int j = 0; j < COLS; j++)
-		{
-			arr[i][j] = rand() % 1000;
-			arr[i][j] /= 100;
-		}
-	}
-}
-void FillRand(char arr[ROWS][COLS], const int ROWS, const int COLS)
-{
-	for (int i = 0; i < ROWS; i++)
-	{
-		for (int j = 0; j < COLS; j++)
-		{
-			arr[i][j] = rand();
-		}
-	}
-}
-template<typename T>void Print(const T arr[], const int n)
-{
-	//Вывод массива на экран:
-	for (int i = 0; i < n; i++)
-	{
-		cout << arr[i] << "\t";
-	}
-	cout << endl;
-}
-template<typename T>void Print(const T arr[ROWS][COLS], const int ROWS, const int COLS)
-{
-	for (int i = 0; i < ROWS; i++)
-	{
-		for (int j = 0; j < COLS; j++)
-		{
-			cout << arr[i][j] << "\t";
-		}
-		cout << endl;
-	}
-}
-template<typename T>T Sum(const T arr[], const int n)
-{
-	//Вычисление суммы элементов массива:
-	T sum = 0;
-	for (int i = 0; i < n; i++)
-	{
-		sum += arr[i];
-	}
-	return sum;
-}
-template<typename T> T Sum(const T arr[ROWS][COLS], const int ROWS, const int COLS)
-{
-	T sum = 0;
-	for (int i = 0; i < ROWS; i++)
-	{
-		for (int j = 0; j < COLS; j++)
-		{
-			sum += arr[i][j];
-		}
-	}
-	return sum;
-}
-template<typename T>double Avg(const T arr[], const int n)
-{
-	return (double)Sum(arr, n) / n;
-}
-template<typename T> double Avg(const T arr[ROWS][COLS], const int ROWS, const int COLS)
-{
-	return (double)Sum(arr, ROWS, COLS) / (ROWS * COLS);
-}
-template<typename T>T minValueIn(const T arr[], const int n)
-{
-	T min = arr[0];
-	for (int i = 0; i < n; i++)
-	{
-		if (arr[i] < min)min = arr[i];
-	}
-	return min;
-}
-template<typename T> T minValueIn(const T arr[ROWS][COLS], const int ROWS, const int COLS)
-{
-	T min = arr[0][0];
-	for (int i = 0; i < ROWS; i++)
-	{
-		for (int j = 0; j < COLS; j++)
-		{
-			if (arr[i][j] < min)min = arr[i][j];
-		}
-	}
-	return min;
-}
-template<typename T>T maxValueIn(const T arr[], const int n)
-{
-	T max = arr[0];
-	for (int i = 0; i < n; i++)
-	{
-		if (arr[i] > max)max = arr[i];
-	}
-	return max;
-}
-template<typename T> T maxValueIn(const T arr[ROWS][COLS], const int ROWS, const int COLS)
-{
-	T max = arr[0][0];
-	for (int i = 0; i < ROWS; i++)
-	{
-		for (int j = 0; j < COLS; j++)
-		{
-			if (arr[i][j] > max)max = arr[i][j];
-		}
-	}
-	return max;
-}
-template<typename T> void Sort(T arr[], const int n)
-{
-	for (int i = 0; i < n; i++)
-	{
-		//Счетчик 'i' выбирает элемент, в который нужно поместить минимальное значение из перебираемых
-		for (int j = i + 1; j < n; j++)
-		{
-			//Счетчик 'j' перебирает элементы, в поисках минимального
-			//arr[i] - выбранный элемент
-			//arr[j] - перебираемый элемент
-			if (arr[j] < arr[i])
-			{
-				T buffer = arr[i];
-				arr[i] = arr[j];
-				arr[j] = buffer;
-			}
-		}
-	}
-}
-template<typename T> void Sort(T arr[ROWS][COLS], const int ROWS, const int COLS)
-{
-	{
-		for (int i = 0; i < ROWS; i++)
-		{
-			for (int j = 0; j < COLS; j++)
-			{
-				for (int k = i; k < ROWS; k++)
-				{
-					for (int l = k == i ? j + 1 : 0; l < COLS; l++)
-					{
-						if (arr[i][j] > arr[k][l])
-						{
-							T tmp = arr[i][j];
-							arr[i][j] = arr[k][l];
-							arr[k][l] = tmp;
-						}
-					}
-				}
-			}
-		}
-	}
-}
-template<typename T>void shiftLeft(T arr[], const int n, int number_of_shifts)
-{
-	for (int i = 0; i < number_of_shifts; i++)
-	{
-		T buffer = arr[0];
-		for (int i = 0; i < n; i++)
-		{
-			arr[i] = arr[i + 1];
-		}
-		arr[n - 1] = buffer;
-	}
-}
-template<typename T> void shiftLeft(T arr[ROWS][COLS], const int ROWS, const int COLS, int number_of_shifts)
-{
-	for (int i = 0; i < number_of_shifts; i++)
-	{
-		int buffer = arr[0][0];
-		for (int i = 0; i < ROWS; i++)
-		{
-			for (int j = 0; j < COLS; j++)
-			{
-				arr[i][j] = arr[i][j + 1];
-			}
-		}
-		arr[ROWS - 1][COLS - 1] = buffer;
-	}
-}
-template<typename T>void shiftRight(T arr[], const int n, int number_of_shifts)
-{
-	for (int i = 0; i < number_of_shifts; i++)
-	{
-		T buffer = arr[n - 1];
-		for (int i = n - 1; i > 0; i--)
-		{
-			arr[i] = arr[i - 1];
-		}
-		arr[0] = buffer;
-	}
-	shiftLeft(arr, n, n - number_of_shifts);
-}
-template<typename T> void shiftRight(T arr[ROWS][COLS], const int ROWS, const int COLS, int number_of_shifts)
-{
-	shiftLeft(arr, ROWS, COLS, ROWS * COLS - number_of_shifts);
-}
-void UniqueRand(int arr[], const int n)
-{
-	for (int i = 0; i < n; i++)
-	{
-		arr[i] = rand() % 10;
-		for (int j = 0; j < i; j++)
-		{
-			if (arr[i] == arr[j])
-			{
-				i--;
-				break;
-			}
-		}
-	}
-}
-void UniqueRand(int arr[ROWS][COLS], const int ROWS, const int COLS)
-{
-	for (int i = 0; i < ROWS; i++)
-	{
-		for (int j = 0; j < COLS; i++)
-		{
-			bool unique;
-			do
-			{
-				arr[i][j] = rand() % (ROWS * COLS);
-				unique = true;
-				for (int k = 0; k <= i; k++)
-				{
-					for (int l = 0; l < (k == i ? j : COLS); l++)
-					{
-						if (arr[i][j] == arr[k][l])
-						{
-							unique = false;
-							break;
-						}
-					}
-					if (!unique)break;
-				}
-			} while (!unique);
-		}
-	}
-}
-void UniqueRand(double arr[ROWS][COLS], const int ROWS, const int COLS)
-{
-	for (int i = 0; i < ROWS; i++)
-	{
-		for (int j = 0; j < COLS; i++)
-		{
-			bool unique;
-			do
-			{
-				arr[i][j] = rand() % (ROWS * COLS);
-				unique = true;
-				for (int k = 0; k <= i; k++)
-				{
-					for (int l = 0; l < (k == i ? j : COLS); l++)
-					{
-						if (arr[i][j] == arr[k][l])
-						{
-							unique = false;
-							break;
-						}
-					}
-					if (!unique)break;
-				}
-			} while (!unique);
-		}
-	}
-}
-void UniqueRand(char arr[ROWS][COLS], const int ROWS, const int COLS)
-{
-	for (int i = 0; i < ROWS; i++)
-	{
-		for (int j = 0; j < COLS; i++)
-		{
-			bool unique;
-			do
-			{
-				arr[i][j] = rand() % (ROWS * COLS);
-				unique = true;
-				for (int k = 0; k <= i; k++)
-				{
-					for (int l = 0; l < (k == i ? j : COLS); l++)
-					{
-						if (arr[i][j] == arr[k][l])
-						{
-							unique = false;
-							break;
-						}
-					}
-					if (!unique)break;
-				}
-			} while (!unique);
-		}
-	}
-}
-template<typename T> void Search(T arr[], const int n)
-{
-	for (int i = 0; i < n; i++)
-	{
-		bool met_before = false;
-		for (int j = 0; j < i; j++)
-		{
-			if (arr[i] == arr[j])
-			{
-				met_before = true;
-				break;
-			}
-		}
-		if (met_before)continue;
-		int count = 0;
-		for (int j = i + 1; j < n; j++)
-		{
-			if (arr[i] == arr[j])
-			{
-				count++;
-			}
-		}
-		if (count)std::cout << "Значение " << arr[i] << " повторяется " << count << " раз" << endl;
-	}
-}
-template<typename T> void Search(T arr[ROWS][COLS], const int ROWS, const int COLS)
-{
-	for (int i = 0; i < ROWS; i++)
-	{
-		for (int j = 0; j < COLS; j++)
-		{
-			bool met_before = false;
-			for (int k = 0; k < ROWS; k++)
-			{
-				for (int l = 0; l < (k == i ? j : COLS); l++)
-				{
-					if (arr[i][j] == arr[k][l])
-					{
-						met_before = true;
-						break;
-					}
-				}
-				if (met_before) break;
-			}
-			if (met_before) continue;
-			int count = 0;
-			for (int k = i; k < ROWS; k++)
-			{
-				for (int l = k == i ? j + 1 : 0; l < COLS; l++)
-				{
-					if (arr[i][j] == arr[k][l])
-					{
-						count++;
-					}
-				}
-			}
-			if (count)std::cout << "Значение " << arr[i][j] << " повторяется " << count << " раз" << endl;
-		}
-	}
 }
