+++ FunctionTemplates/stdafx.h
@@ -1,0 +1,7 @@
+#pragma once
+#include<iostream>
+using namespace std;
+
+using std::cin;
+using std::cout;
+using std::endl;
